Player class  생성
