+++ src/main/java/repository/Computer.java
@@ -1,22 +1,9 @@
 package repository;
 
-import java.util.LinkedHashSet;
-import java.util.Set;
-
-public class Computer {
+public class Computer extends Player {
 
 	public int generate() {
 		return (int) (Math.random() * 10);
 	}
 
-	public Set<Integer> save() {
-		Set<Integer> callNumbers = new LinkedHashSet<>();
-
-		while (callNumbers.size() != 3) {
-			int number = generate();
-			callNumbers.add(number);
-		}
-
-		return callNumbers;
-	}
 }
